Add Policy Logic
Add policy commands, utils, logic and ui

There is a need to have view and remove commands for policy for a MVP.
The Ui and logic that supports viewing and removing policies improve
user experience.

Let's add the RemovePolicy and ViewPolicy Commands, as well as update
the storage for the addressbook and add a storage for policy. Let's also
add policy related method into the util, as well as create a PolicyCard
and PolicyListPanel in Ui to display policies. Let's provide an early
iteration of our implementation for our teamates to refer to and work on
their portion of code, given they may cross reference to policy to
implement more MVP features.