Add ViewPolicy Command, Parser, Test
Add View Policy Command, Parser and Test

There is a need for the view policy command to view all policies in the
address book. There is a need for a parser to parse the user input to
show all policies. Testcases validate that these view policies code work

Let's add the view policy command, parser, and a view policy command
test to the addressbook