Fix Allow Duplicate Prefix and Alignment
- Fix bug where duplicate prefixes are allowed in view commands
- Refactor ViewContact code to align with other view commands and require
prefix to specify specific contact to view
- Refactor static string for Appointment to use
  MESSAGE_INVALID_DATE_FORMAT instead for consistency

The user should not be able to key in two duplicate prefixes, where the
current implementation would only accept the last prefix. There is a
need to resolve this bug so the code works as intended.

Let's add a verifynoduplicate method to each viewcommandparser, and
refactor the viewcontactcommand.