Add EditAppointment logic
Add EditAppointmentCommand and Parser.
Update AddressBookParser to include edit appointment.

There is a need to allow users to edit the appointments in case there
are any changes to be made. This removes the need to remove and add a
new appointment just to edit details.

Let's add the EditAppointmentCommand and its Parser to the addressbook.