Add Testcases for EditAppointmentCommand
Add testcases for EditAppointmentCommand and
EditAppointmentCommandParser.
Add EditAppointmentDescriptor and Builder Tests to support testing for
editappointments.

There is a need to test for functionality as to whether editappointment
will work for different user inputs

Let's add testcases for editappointment, and build helper methods and
classes like editappointmentdescriptorbuilder to test the functionality
of the editappointment command and parser.