Fix FilteredList Logic for Commands
Refactor RemoveAppointment, RemovePolicy, EditAppointment and EditPolicy
to use hasAppointment or hasPolicy to verify existence in addressbook,
and getAppointment getPolicy to get the Object from addressbook.

There is a need for the update to fix the way user view the pages. User
could only initially update / remove whatever was on the page. Since the
pages are dynamic now, users should be allowed to remove/edit from a
different page if necessary, at least for Appointments and Policies
where unique identifier is used instead of page list index for Person