Tighten duplicate contract logic
Previously contracts were not different as long as any field is different.
That is name and ContractId does not affect duplicate contract checks.
But as long as any of the below fields are different, even if the rest
are identical, the contract is different:
* Nric
* PolicyId
* Date Signed
* Expiry Date
* Premium

Now contracts are duplicate as long Nric and PolicyId is the same